Extracted StackConsumer to own file and added documentation
